More questions, answers and tips!
Big update. Added two more questions, A TON of more answers so now the user doesn't have to repeat the whole quiz if they respond in a way the program didn't think about. Also, tips! The user can sacrifice a score point for a tip on the next answers, when the tip isn't useful anymore, the user will get notified. Also!! I have done the ranks, but I have them on my school computer, where we program only on Wednesdays, so expect an update on Wednesday with all the ranks you can get! ♡♡