Ranks!!
As promised, the ranks have been added, as I add more questions, the ranks will have a ratio of score points (ex: Silver 1 will be 1-3 points to get It, Silver 2 4-6...etc) to get them instead of just a unique number. I have also added lists for the upcoming answers and have credited the page I use for the ranks aswell as changing the version from 0.5 to 0.7.
I am getting closer and closer to getting the quiz finished, and I can't wait! I love this. By the way, when I finish the quiz I will start working on the UI of course. Also something that I wanted to do was making It so that the questions are random, I will look into that too.
What I expect is:
v0.8 - All questions & answers.
v0.9 - Random questions + misc changes maybe.
v1.0 - Actual application with UI.
But this may change, this is only my plans.